third_party/blink/web_tests/external/wpt/mediacapture-streams/MediaStream-default-feature-policy.https-expected.txt

This is a testharness.js-based test.
[FAIL] Feature policy "microphone" can be enabled in cross-origin iframes using "allow" attribute.
  assert_equals: expected "#OK" but got "#All requested devices must be present with audio true and video true, or fail"
[FAIL] Feature policy "camera" can be enabled in cross-origin iframes using "allow" attribute.
  assert_equals: expected "#OK" but got "#All requested devices must be present with audio true and video true, or fail"
Harness: the test ran to completion.
